requirements.txt
pytorch
pandas
numpy
scikit-learn
tqdm
matplotlib
seaborn
jupyter
